requirements.txt
absl-py==0.13.0
av
diffusers==0.11.1
dill==0.3.5.1
dm_control
easydict
einops==0.4.1
gpustat
h5py
hydra-core==1.2.0
imageio
imageio-ffmpeg
ipdb
jupyterlab==3.0.14
kaleido
matplotlib
mediapy
moviepy
natsort
notebook
numba==0.56.4
opencv-python==4.5.3.56
open3d
pandas
plotly
psutil
pygame==2.1.2
pyparsing==2.4.7
pyrealsense2==2.54.2.5684
pyspacemouse==1.1.5
pynput==1.8.1
pymodbus==2.1.0
scipy==1.10.1
scikit-learn==1.3.2
scikit-image
scikit-video
numcodecs==0.12.1
numpy==1.24.4
omegaconf==2.3.0
huggingface_hub==0.36.2
safetensors==0.5.3
setuptools==59.5.0
shapely==1.8.4
spnav==0.9
ur_rtde==1.6.1
tensorboard==2.10.1
tensorboardx==2.5.1
termcolor
timm==0.9.7
torch_geometric
tqdm==4.67.3
tyro
visdom
wandb
yapf==0.31.0
zarr==2.12.0
